improve exit code types for systems that support it

+++ src/cli.rs
@@ -2,12 +2,12 @@ use std::env;
 use std::process::exit;
 
 pub struct ParsedArgs {
-    pub exit_code: u8,
+    pub exit_code: i32,
 }
 
 pub fn parse_args() -> ParsedArgs {
     if let Some(code_str) = env::args().nth(1) {
-        let exit_code = code_str.parse::<u8>()
+        let exit_code = code_str.parse::<i32>()
             .unwrap_or_else(|_err| {
                 eprintln!("blazesh: fialed to parse exit code argument");
                 exit(1);
